Remove privacy status tests and add tests for events queing with consent
