+++ Podfile
@@ -24,6 +24,7 @@ target 'FunctionalTests' do
   pod 'AEPServices', :git => 'https://github.com/adobe/aepsdk-core-ios.git', :branch => 'dev-v3.0.1'
   pod 'AEPRulesEngine'
   pod 'AEPIdentityEdge', :git => 'git@github.com:adobe/aepsdk-identityedge-ios.git', :branch => 'dev'
+  pod 'AEPConsent', :git => 'git@github.com:adobe/aepsdk-consentedge-ios.git', :branch => 'dev'
 end
 
 target 'AEPDemoAppSwiftUI' do
